feat: add Loader by skeleton
