+++ client/README.md
@@ -10,4 +10,5 @@ $ npm i react-icons
 $ npm i js-cookie
 $ npm i --save-dev @types/js-cookie
 $ npm i swiper
+$ npm install react-loading-skeleton
 ```
