+++ client/components/ui/Loader.tsx
@@ -1,0 +1,9 @@
+import React, { FC } from 'react';
+import Skeleton, { SkeletonProps } from 'react-loading-skeleton';
+import 'react-loading-skeleton/dist/skeleton.css';
+
+const Loader: FC<SkeletonProps> = props => {
+	return <Skeleton {...props} />;
+};
+
+export default Loader;
