--- client/components/ui/layout/sidebar/profile-info/useProfile.ts
@@ -1,1 +1,0 @@
-export const useProfile = () => {};
