+++ client/package.json
@@ -24,6 +24,7 @@
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.9",
     "react-icons": "^4.8.0",
+    "react-loading-skeleton": "^3.3.1",
     "react-query": "^3.39.3",
     "sass": "^1.62.1",
     "swiper": "^9.3.2",
